+++ cfg.cfg
@@ -15,6 +15,8 @@ time_server = "192.168.0.89"
   arduino_linear = [ 'H','T']
   # list of sensors, which values can vary exponentially, to be requested to Arduino
   arduino_exp = [ 'R','G','B','C']
+  # Bosh barometric pressure BMP180
+  bmp = true
   # sensor adjustments
   [ sensors.adj_H ]
       value = +10
@@ -23,7 +25,7 @@ time_server = "192.168.0.89"
 
 # Arduino Serial setup
 [serial]
-  tty = "/dev/ttyUSB0"
+  tty = "none"
   baud = 9600
   timeout = 1950 #in milliseconds
 
@@ -31,8 +33,8 @@ time_server = "192.168.0.89"
 [ inputs ]
   [ inputs.human ]
      pin = 0
-   [ inputs.bmp180 ]
-     pin = 0
+   [inputs.rain]
+     pin = 4
 
 #  GPIO Outputs
 [ outputs ]
@@ -41,7 +43,7 @@ time_server = "192.168.0.89"
   [ outputs.socket1 ]
     pin = 0
   [ outputs.socket2 ]
-    pin = 14
+    pin = 0
 
 [ data_analysis ]
   # historic data depth: how many samples we keep to estimate a good value
@@ -74,15 +76,15 @@ time_server = "192.168.0.89"
   token = "xoxp-513737629045-514938072262-514943142246-2b50efd8e3f1fe21a4....."
 
 [ lighting ]
-# Remember to set GPIO2 Pin
+# Remember to set GPIO2 Pin to command the light relays
   [lighting.morning_start]
-    hour = 7
+    hour = 0
   [lighting.morning_end]
-    hour = 9
+    hour = 0
   [lighting.evening_start]
-    hour  = 16
+    hour = 0
   [lighting.evening_end]
-    hour  = 22
+    hour = 0
 
 [ speech ]
   # first words used before the stats
