+++ config.go
@@ -37,6 +37,7 @@ type sensors struct {
 	Poll_interval  int      `toml:"poll_interval"`
 	Adj_H          map[string]int
 	Adj_T          map[string]int
+	Bmp	       bool	`toml:"bmp"`
 }
 
 type speech struct {
